Cambios tracking - etalent nueva version

+++ purchaseTracking/Models/eTALENT/HISTORICO_VACACIONES.cs
@@ -11,12 +11,11 @@ namespace purchaseTracking.Models.eTALENT
         public string NOMBRE_COMPLETO { get; set; }
         public string FECHA_INGRESO { get; set; }
         public string PERIODO { get; set; }
-        public int DIAS { get; set; }
-        public int GOZADOS { get; set; }
-        public int DVCON_DIAS { get; set; }
+        public Single DIAS { get; set; }
+        public Single GOZADOS { get; set; }
+        public Single DVCON_DIAS { get; set; }
         public string DESDE { get; set; }
         public string HASTA { get; set; }
         public string CODIGO_EXTERNO { get; set; }
-
     }
 }
